refactor : 프로젝트 비즈니스 로직을 읽기 전용 비즈니스 로직으로 분리함
